+++ src/StringEncoding/InvalidJSONException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Ingenerator\PHPUtils\StringEncoding;
+
+class InvalidJSONException extends \InvalidArgumentException
+{
+
+}